+++ Source/Anvil/Models/CommonConfigKeys.cs
@@ -1,0 +1,9 @@
+﻿namespace Anvil.Models
+{
+    public sealed class CommonConfigKeys
+    {
+        public const string CloseToSystemTray = "CloseToSystemTray";
+        public const string IncludePreRelease = "IncludePreRelease";
+        public const string UpdateUrl = "UpdateUrl";
+    }
+}