+++ Source/Anvil/Services/ServiceModule.cs
@@ -1,4 +1,6 @@
-﻿using Autofac;
+﻿using Anvil.Models;
+
+using Autofac;
 
 using Squirrel;
 
@@ -6,6 +8,8 @@ namespace Anvil.Services
 {
     public sealed class ServiceModule : Module
     {
+        private const string DEFAULT_UPDATE_URL = "https://github.com/garrettpauls/Anvil";
+
         protected override void Load(ContainerBuilder builder)
         {
             var assemblies = new[] {typeof(ServiceModule).Assembly};
@@ -27,10 +31,13 @@ namespace Anvil.Services
             builder
                 .Register(context =>
                 {
-                    var config = context.Resolve<IConfiguration>();
+                    var config = context.Resolve<IConfigurationService>();
+                    var updateUrl = config.GetValue(CommonConfigKeys.UpdateUrl, () => DEFAULT_UPDATE_URL);
+                    var includePreRelease = config.GetValue(CommonConfigKeys.IncludePreRelease, () => false);
+
                     return UpdateManager.GitHubUpdateManager(
-                        config.UpdateUrl,
-                        prerelease: config.IncludePreReleaseVersions).Result;
+                        updateUrl,
+                        prerelease: includePreRelease).Result;
                 })
                 .As<IUpdateManager>()
                 .SingleInstance();