+++ Source/Anvil/Views/ConfigurationUI/SettingsViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using Anvil.Framework.MVVM;
+using Anvil.Models;
 using Anvil.Services;
 
 using ReactiveUI;
@@ -7,12 +8,32 @@ namespace Anvil.Views.ConfigurationUI
 {
     public sealed class SettingsViewModel : RoutableViewModel
     {
-        public SettingsViewModel(IConfiguration configuration, IScreen hostScreen)
+        private readonly IConfigurationService mConfiguration;
+
+        public SettingsViewModel(IConfigurationService configuration, IScreen hostScreen)
             : base("settings", hostScreen)
         {
-            Configuration = configuration;
+            mConfiguration = configuration;
         }
 
-        public IConfiguration Configuration { get; }
+        public bool CloseToSystemTray
+        {
+            get { return mConfiguration.GetValue(CommonConfigKeys.CloseToSystemTray, () => false); }
+            set
+            {
+                mConfiguration.SetValue(CommonConfigKeys.CloseToSystemTray, value);
+                this.RaisePropertyChanged();
+            }
+        }
+
+        public bool IncludePreReleaseVersions
+        {
+            get { return mConfiguration.GetValue(CommonConfigKeys.IncludePreRelease, () => false); }
+            set
+            {
+                mConfiguration.SetValue(CommonConfigKeys.IncludePreRelease, value);
+                this.RaisePropertyChanged();
+            }
+        }
     }
 }